+++ lib/group_avatar.dart
@@ -1,3 +1,3 @@
 library chat_group_avatar;
 
-export 'src/group_avatar.dart';
+export 'src/group_avatar.dart' show GroupAvatar, AvatarShape, GroupAvatarLayout;